# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InterfaceConfirmationRequest": {
        "properties": {
          "access_point": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Access Point",
            "type": "string"
          },
          "wan": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Wan",
            "type": "string"
          }
        },
        "required": [
          "wan",
          "access_point"
        ],
        "title": "InterfaceConfirmationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OnionPi API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/onboarding/clock": {
      "post": {
        "operationId": "onboarding_clock_api_v1_onboarding_clock_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboarding Clock Api V1 Onboarding Clock Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboarding Clock"
      }
    },
    "/api/v1/onboarding/firewall": {
      "post": {
        "operationId": "onboarding_firewall_api_v1_onboarding_firewall_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboarding Firewall Api V1 Onboarding Firewall Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboarding Firewall"
      }
    },
    "/api/v1/onboarding/interfaces": {
      "post": {
        "operationId": "onboarding_interfaces_api_v1_onboarding_interfaces_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboarding Interfaces Api V1 Onboarding Interfaces Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboarding Interfaces"
      }
    },
    "/api/v1/onboarding/recovery-code": {
      "post": {
        "operationId": "onboarding_recovery_code_api_v1_onboarding_recovery_code_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboarding Recovery Code Api V1 Onboarding Recovery Code Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboarding Recovery Code"
      }
    },
    "/api/v1/onboarding/recovery-code/confirm": {
      "post": {
        "operationId": "onboarding_recovery_code_confirm_api_v1_onboarding_recovery_code_confirm_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboarding Recovery Code Confirm Api V1 Onboarding Recovery Code Confirm Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboarding Recovery Code Confirm"
      }
    },
...
  }
}